+++ src/pkl/mod.rs
@@ -1,4 +1,3 @@
-pub mod datasize;
 mod de;
 mod deserializer;
 pub(crate) mod internal;
@@ -7,6 +6,6 @@ mod serializer;
 
 pub use deserializer::Deserializer;
 pub use internal::IntSeq;
-pub use internal::PklValue;
+
 pub(crate) use pkl_mod::PklMod;
 pub use serializer::PklSerialize;
